Move `get_provisioning_timeout`
